+++ sdk/resourcemanager/recoveryservices/armrecoveryservicessiterecovery/replicationappliances_client.go
@@ -45,7 +45,7 @@ func NewReplicationAppliancesClient(subscriptionID string, credential azcore.Tok
 
 // NewListPager - Gets the list of Azure Site Recovery appliances for the vault.
 //
-// Generated from API version 2023-08-01
+// Generated from API version 2024-01-01
 //   - resourceName - The name of the recovery services vault.
 //   - resourceGroupName - The name of the resource group where the recovery services vault is present.
 //   - options - ReplicationAppliancesClientListOptions contains the optional parameters for the ReplicationAppliancesClient.NewListPager
@@ -96,7 +96,7 @@ func (client *ReplicationAppliancesClient) listCreateRequest(ctx context.Context
 	if options != nil && options.Filter != nil {
 		reqQP.Set("$filter", *options.Filter)
 	}
-	reqQP.Set("api-version", "2023-08-01")
+	reqQP.Set("api-version", "2024-01-01")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil